Cleaning up in probe.js and created mp4/captions-parser.js: - Use existing timescale() method. - Only use video track timescales, parse the segment only once rather than for each video track. - Moving all caption parsing methods to new file captions-parser.js
