+++ lib/index.js
@@ -4,8 +4,7 @@ var muxjs = {
   codecs: require('./codecs'),
   mp4: require('./mp4'),
   flv: require('./flv'),
-  mp2t: require('./m2ts'),
-  captionsParser: require('./m2ts/captions-parser')
+  mp2t: require('./m2ts')
 };
 
 // include all the tools when the full library is required
