+++ lib/mp4/index.js
@@ -3,5 +3,5 @@ module.exports = {
   Transmuxer: require('./transmuxer').Transmuxer,
   AudioSegmentStream: require('./transmuxer').AudioSegmentStream,
   VideoSegmentStream: require('./transmuxer').VideoSegmentStream,
-  probe: require('./probe')
+  captionsParser: require('./captions-parser')
 };
