+++ lib/mp4/probe.js
@@ -9,12 +9,7 @@
 'use strict';
 
 var toUnsigned = require('../utils/bin').toUnsigned;
-var discardEmulationPreventionBytes = require('../m2ts/captions-parser').discardEmulationPreventionBytes;
-var CaptionStream = require('../m2ts/caption-stream').CaptionStream;
-var findBox, parseType, timescale, startTime,
-  findSeiNals, mapToSample, parseSamples,
-  parseHandlerType, parseTrackId, parseMdhd, parseDecodeTime, parseTfhd,
-  parseInitForCaptionMetadata, parseCaptionNals, parseEmbeddedCaptions;
+var findBox, parseType, timescale, startTime;
 
 // Find the data for a box specified by its path
 findBox = function(data, path) {
@@ -187,469 +182,9 @@ startTime = function(timescale, fragment) {
   return isFinite(result) ? result : 0;
 };
 
-mapToSample = function(offset, samples) {
-  for (var i = 0; i < samples.length - 1; i++) {
-    var sample = samples[i];
-    var nextSample = samples[i + 1];
-
-    if (sample.dataOffset <= offset &&
-        offset < nextSample.dataOffset) {
-      return sample;
-    }
-  }
-
-  return null;
-};
-
-findSeiNals = function(avcStream, samples, trackId) {
-  var
-    avcView = new DataView(avcStream.buffer, avcStream.byteOffset, avcStream.byteLength),
-    result = [],
-    seiNal,
-    i,
-    length;
-
-  for (i = 0; i + 4 < avcStream.length; i += length) {
-    length = avcView.getUint32(i);
-    i += 4;
-
-    // bail if this doesn't appear to be an H264 stream
-    if (length <= 0) {
-      continue;
-    }
-
-    switch (avcStream[i] & 0x1F) {
-    case 0x06:
-      var data = avcStream.subarray(i + 1, i + 1 + length);
-      seiNal = {
-        nalUnitType: 'sei_rbsp',
-        size: length,
-        data: data,
-        escapedRBSP: discardEmulationPreventionBytes(data),
-        trackId: trackId
-      };
-      var matchingSample = mapToSample(i, samples);
-      if (matchingSample) {
-        seiNal.pts = matchingSample.pts;
-        seiNal.dts = matchingSample.dts;
-      } else {
-        seiNal.pts = 0;
-        seiNal.dts = 0;
-      }
-      result.push(seiNal);
-      break;
-    default:
-      break;
-    }
-  }
-
-  return result;
-};
-
-parseMdhd = function(bytes) {
-  var view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
-  var version = view.getUint8(0);
-  var creationTime;
-  var modificationTime;
-  var timescale;
-  var duration;
-
-  if (version === 1) {
-    // 64 bits
-    creationTime = bytes.subarray(4, 12);
-    modificationTime = bytes.subarray(12, 20);
-    // 32 bits
-    timescale = view.getUint32(20);
-    // 64 bits
-    duration = bytes.subarray(24, 32);
-
-  } else {
-    // 32 bits
-    creationTime = view.getUint32(4);
-    modificationTime = view.getUint32(8);
-    timescale = view.getUint32(12);
-    duration = view.getUint32(16);
-  }
-
-  return {
-    creationTime: creationTime,
-    modificationTime: modificationTime,
-    timescale: timescale,
-    duration: duration
-  };
-};
-
-parseTrackId = function(bytes) {
-  var view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
-  var version = view.getUint8(0);
-  var i = (version === 0) ? 12 : 20;
-  var trackId = view.getUint32(i);
-
-  return trackId;
-};
-
-parseHandlerType = function(bytes) {
-  // 32 bits, skip the version, flags, predefined 0
-  var handlerType = parseType(bytes.subarray(8, 12));
-  var name = '';
-
-  // skip reserved integers
-  for (var i = 24; i < bytes.length; i++) {
-    // found null byte
-    if (bytes[i] === 0x00) {
-      break;
-    }
-    name += String.fromCharCode(bytes[i]);
-  }
-
-  // FIXME: I don't think name is right for the label
-
-  return handlerType;
-};
-
-parseInitForCaptionMetadata = function(initSegment) {
-  var traks = findBox(initSegment, ['moov', 'trak']);
-  var result = {};
-
-  for (var i = 0; i < traks.length; i++) {
-    var trak = traks[i];
-    var trakResult;
-
-    var mdia = findBox(trak, ['mdia']);
-    var tkhd = findBox(trak, ['tkhd']);
-    var trackId = parseTrackId(tkhd[0]);
-
-    // 1 per trak
-    var hdlr = findBox(mdia[0], ['hdlr']);
-    var handlerType = parseHandlerType(hdlr[0]);
-
-    // Don't return audio track metadata
-    if (handlerType !== 'vide') {
-      continue;
-    }
-
-    var stbl = findBox(mdia[0], ['minf', 'stbl']);
-    var mdhd = findBox(mdia[0], ['mdhd']);
-
-    var parsedMdhd = parseMdhd(mdhd[0]);
-
-    trakResult = {
-      trackId: trackId,
-      timescale: parsedMdhd.timescale,
-      duration: parsedMdhd.duration,
-      handlerType: handlerType,
-      stbl: stbl[0]
-    };
-
-    result[trackId] = trakResult;
-  }
-
-  return result;
-};
-
-parseDecodeTime = function(bytes) {
-  var view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
-  var version = view.getUint8(0);
-  // An  integer  equal  to  the  sum  of  the  decode  durations
-  // of  all  earlier  samples in the media, expressed in the
-  // media's timescale. It does not include the samples added
-  // in the enclosing track fragment.
-  var baseMediaDecodeTime = view.getUint32(4);
-
-  // FIXME: double check if this is correct?
-  if (version === 1) {
-    baseMediaDecodeTime *= Math.pow(2, 32);
-  }
-
-  return baseMediaDecodeTime;
-};
-
-parseTfhd = function(bytes) {
-  var result = {};
-  var view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
-  var flags;
-  // Parsed Flags
-  var baseDataOffsetPresent;
-  var sampleDescriptionsIndexPresent;
-  var defaultSampleDurationPresent;
-  var defaultSampleSizePresent;
-  var defaultSampleFlagsPresent;
-  var durationIsEmpty;
-  var defaultBaseIsMoof;
-  var i;
-
-  // 24 bits, skip version
-  flags = bytes.subarray(1, 4);
-  // 32 bits
-  result.trackId = view.getUint32(4);
-
-  baseDataOffsetPresent = flags[2] & 0x01;
-  sampleDescriptionsIndexPresent = flags[2] & 0x02;
-  defaultSampleDurationPresent = flags[2] & 0x08;
-  defaultSampleSizePresent = flags[2] & 0x10;
-  defaultSampleFlagsPresent = flags[2] & 0x20;
-  // FIXME: Do we actually need these two?
-  durationIsEmpty = flags[0] & 0x010000;
-  defaultBaseIsMoof =  flags[0] & 0x020000;
-  i = 8;
-
-  // These are optional values
-  if (baseDataOffsetPresent) {
-    // 64 bits
-    result.baseDataOffset = bytes.subarray(i, i + 8);
-    i += 8;
-  }
-
-  if (sampleDescriptionsIndexPresent) {
-    // 32 bits
-    result.sampleDescriptionsIndex = view.getUint32(i);
-    i += 4;
-  }
-
-  if (defaultSampleDurationPresent) {
-    result.defaultSampleDuration = view.getUint32(i);
-    i += 4;
-  }
-
-  if (defaultSampleSizePresent) {
-    result.defaultSampleSize = view.getUint32(i);
-    i += 4;
-  }
-
-  if (defaultSampleFlagsPresent) {
-    result.defaultSampleFlags = view.getUint32(i);
-  }
-
-  return result;
-};
-
-parseSamples = function(bytes, baseMediaDecodeTime, tfhd, timescale) {
-  var view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
-  var samples = [];
-
-  // Default values for samples
-  var sampleDataOffset = tfhd.baseDataOffset || 0; // FIXME: set to baseDataOffset (from moof and tfhd)
-  // FIXME: need to actually use these as defaults now
-  var defaultSampleDuration = tfhd.defaultSampleDuration;
-  var defaultSampleSize = tfhd.defaultSampleSize;
-  var defaultSampleFlags = tfhd.defaultSampleFlags;
-  var trackId = tfhd.trackId;
-
-  // Initialize decode time stamp
-  var currentDts = baseMediaDecodeTime;
-  var version;
-  var sampleCount;
-  var flags;
-
-  // Top Level Flags for Optional Fields
-  var dataOffsetPresent;
-  var firstSampleFlagsPresent;
-  var sampleDurationPresent;
-  var sampleSizePresent;
-  var sampleFlagsPresent;
-  var sampleCompositionTimeOffsetPresent;
-  var i;
-
-  version = view.getUint8(0);
-  flags = bytes.subarray(1, 4);
-  // 32 bits
-  sampleCount = view.getUint32(4);
-  i = 8;
-
-  // Flag interpretation
-  dataOffsetPresent = flags[2] & 0x01;
-  firstSampleFlagsPresent = flags[2] & 0x04;
-  // Comparing with 2nd byte of 0x100, 0x200, 0x400, 0x800
-  sampleDurationPresent = flags[1] & 0x01;
-  sampleSizePresent = flags[1] & 0x02;
-  sampleFlagsPresent = flags[1] & 0x04;
-  sampleCompositionTimeOffsetPresent = flags[1] & 0x08;
-
-  /**
-    * Optional Fields
-   **/
-
-  if (dataOffsetPresent) {
-    // 32 bit signed integer
-    sampleDataOffset += view.getInt32(i);
-    i += 4;
-  }
-
-  // Overrides the flags for the first sample only
-  // The order of optional values will be:
-  //   duration, size, compositionTimeOffset
-  if (firstSampleFlagsPresent) {
-    var firstSampleDuration;
-    var firstSampleSize;
-    var firstSampleCompositionTimeOffset;
-
-    var firstSampleFlags = bytes.subarray(i, i + 4);
-    i += 4;
-
-    if (sampleDurationPresent) {
-      firstSampleDuration = view.getUint32(i);
-      i += 4;
-    }
-
-    if (sampleSizePresent) {
-      firstSampleSize = view.getUint32(i);
-      i += 4;
-    }
-
-    if (sampleCompositionTimeOffsetPresent) {
-      firstSampleCompositionTimeOffset = (version === 0) ?
-        view.getUint32(i) : view.getInt32(i);
-      i += 4;
-    }
-
-    samples.push({
-      trackId: trackId,
-      duration: firstSampleDuration,
-      size: firstSampleSize,
-      flags: firstSampleFlags,
-      compositionTimeOffset: firstSampleCompositionTimeOffset,
-      dataOffset: sampleDataOffset,
-      dts: currentDts,
-      pts: currentDts + firstSampleCompositionTimeOffset
-    });
-
-    sampleDataOffset += firstSampleSize;
-    currentDts += firstSampleDuration;
-  }
-
-  for (var j = 0; j < sampleCount; j++) {
-    var sampleDuration;
-    var sampleSize;
-    var sampleFlags;
-    var sampleCompositionTimeOffset;
-
-    if (sampleDurationPresent) {
-      sampleDuration = view.getUint32(i);
-      i += 4;
-    }
-
-    if (sampleSizePresent) {
-      sampleSize = view.getUint32(i);
-      i += 4;
-    }
-
-    if (sampleFlagsPresent) {
-      sampleFlags = bytes.subarray(i, i + 4);
-      i += 4;
-    }
-
-    if (sampleCompositionTimeOffsetPresent) {
-      sampleCompositionTimeOffset = (version === 0) ?
-        view.getUint32(i) : view.getInt32(i);
-      i += 4;
-    }
-
-    samples.push({
-      trackId: trackId,
-      duration: sampleDuration,
-      size: sampleSize,
-      flags: sampleFlags,
-      compositionTimeOffset: sampleCompositionTimeOffset,
-      dataOffset: sampleDataOffset,
-      dts: currentDts,
-      pts: currentDts + sampleCompositionTimeOffset
-    });
-
-    // Update data offset, dts
-    sampleDataOffset += sampleSize;
-    currentDts += sampleDuration;
-  }
-
-  return samples;
-};
-
-parseCaptionNals = function(segment, timescale) {
-  // Samples
-  var trafs = findBox(segment, ['moof', 'traf']);
-  var sampleInfo = {};
-  // SEI NAL units
-  var mdats = findBox(segment, ['mdat']);
-  var captionNals = [];
-
-  // Parse moofs for the sample table
-  trafs.forEach(function(traf, index) {
-    var trafResult;
-
-    var tfhd = findBox(traf, ['tfhd']);
-    var tfdt = findBox(traf, ['tfdt']);
-    var trun = findBox(traf, ['trun']);
-    var subs = findBox(traf, ['subs']);
-
-    var headerInfo = parseTfhd(tfhd[0]);
-    var baseMediaDecodeTime = parseDecodeTime(tfdt[0]);
-
-    trafResult = {
-      trackId: headerInfo.trackId,
-      samples: parseSamples(trun[0], baseMediaDecodeTime, headerInfo, timescale),
-      subSamples: subs
-    };
-
-    sampleInfo[index] = trafResult;
-  });
-
-  // Parse mdats for SEI messages
-  mdats.forEach(function(mdat, index) {
-    // mdat and moof paired by index
-    var samples = sampleInfo[index].samples;
-    var trackId = sampleInfo[index].trackId;
-    var seiNals = findSeiNals(mdat, samples, trackId);
-
-    captionNals = captionNals.concat(seiNals);
-  });
-
-  return captionNals;
-};
-
-/**
- * Parses out inband captions from an MP4 container
- *
- * @param {TypedArray} init The init segment
- * @param {TypedArray} segment The fmp4 segment containing imbedded captions
- **/
-parseEmbeddedCaptions = function(init, segment) {
-  // FIXME: should each trak have it's own caption stream?
-  var captionStreams = [];
-  var parsedCaptions = [];
-  // Parse out media header information for ISOBMFF video tracks
-  var initInfo = parseInitForCaptionMetadata(init);
-
-  // Parse out 608/708 caption NAL units from segment
-  for (var trackId in initInfo) {
-    var trakInfo = initInfo[trackId];
-    var timescale = trakInfo.timescale;
-    var seiNals = parseCaptionNals(segment, timescale);
-    var captionStream = new CaptionStream();
-
-    // Collect dispatched captions
-    captionStream.on('data', function(event) {
-      // Convert to seconds
-      event.startTime = event.startPts / timescale;
-      event.endTime = event.endPts / timescale;
-
-      parsedCaptions.push(event);
-    });
-
-    seiNals.forEach(function(nal) {
-      captionStream.push(nal);
-    });
-
-    // Force the parsed captions to be dispatched
-    captionStream.flush();
-
-    captionStreams[trackId] = captionStream;
-  }
-
-  return parsedCaptions;
-};
-
 module.exports = {
+  findBox: findBox,
   parseType: parseType,
   timescale: timescale,
-  startTime: startTime,
-  parseEmbeddedCaptions: parseEmbeddedCaptions
+  startTime: startTime
 };
